% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dose_update_numbers_from_scheme.R
\name{adjust_dose_checks}
\alias{adjust_dose_checks}
\title{Checks for dose_update_number obtained from dose_update scheme}
\usage{
adjust_dose_checks(adjust_at_dose, regimen)
}
\description{
Checks for dose_update_number obtained from dose_update scheme
}
