% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_sim_output.R
\name{bind_sim_output}
\alias{bind_sim_output}
\title{Bind simulated subject-level output together
(from list into data.frames)}
\usage{
bind_sim_output(res)
}
\arguments{
\item{res}{temporary result object (list) in \code{run_trial()}}
}
\description{
Bind simulated subject-level output together
(from list into data.frames)
}
