% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_metrics.R
\name{calc_auc_from_regimen}
\alias{calc_auc_from_regimen}
\title{Get AUC from a regimen}
\usage{
calc_auc_from_regimen(regimen, parameters, model, t_obs, ...)
}
\arguments{
\item{regimen}{PKPDsim regimen object}

\item{parameters}{use MAP estimation to get estimated AUC, use true patient
parameters to get true AUC. Parameters must correspond to the model used.
Accepts parameters supplied as a data frame row, a named vector or as a
list.}

\item{model}{model to use for AUC calculations.}

\item{t_obs}{timepoint(s) at which to calculate AUC}

\item{...}{arguments passed on to PKPDsim::sim. Typical arguments include
\code{covariates} or \code{iov_bins}}
}
\value{
numeric vector of AUCs between each simulated time point. Control
time period over which AUC should be calculated using \code{t_obs}.
}
\description{
Supply the final regimen and the final parameter estimates to get the final
estimated AUC. Supply the final regimen and the true individual parameter
estimates to get the final true AUC.
}
