% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_metrics.R
\name{calc_auc_from_sim}
\alias{calc_auc_from_sim}
\title{Get AUC from a simulation}
\usage{
calc_auc_from_sim(sim_output, auc_comp)
}
\arguments{
\item{sim_output}{output of a \code{PKPDsim::sim} call}

\item{auc_comp}{auc compartment}
}
\value{
numeric vector of AUCs between each simulated time point. Control
time period over which AUC should be calculated using \code{t_obs} argument to
\code{PKPDsim::sim}.
}
\description{
Get AUC from a simulation
}
