% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_metrics.R
\name{calc_concentration_from_regimen}
\alias{calc_concentration_from_regimen}
\title{Get concentration from a regimen}
\usage{
calc_concentration_from_regimen(regimen, parameters, model, target_design, ...)
}
\arguments{
\item{regimen}{PKPDsim regimen object}

\item{parameters}{use MAP estimation to get estimated concentration, use true
patient parameters to get true concentration. Parameters must correspond to
the model used. Accepts parameters supplied as a data frame row, a named
vector or as a list.}

\item{model}{model to use for concentration calculations.}

\item{target_design}{target design, created using \code{create_target_design()}}

\item{...}{arguments passed on to PKPDsim::sim. Typical arguments include
\code{covariates} or \code{iov_bins}}
}
\value{
numeric vector of concentrations between each simulated time point.
Control when concentration is estimated using \code{target_time}.
}
\description{
Supply the final regimen and the final parameter estimates to get the final
estimated AUC. Supply the final regimen and the true individual parameter
estimates to get the final true AUC.
}
