% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_metrics.R
\name{calc_time_to_target}
\alias{calc_time_to_target}
\title{Calculate time to target attainment}
\usage{
calc_time_to_target(
  regimen,
  target_design,
  auc_comp,
  model,
  parameters,
  covariates,
  ...
)
}
\arguments{
\item{regimen}{PKPDsim regimen object}

\item{target_design}{target design, created using \code{create_target_design()}}

\item{auc_comp}{compartment to look for AUC}

\item{model}{PKPDsim model object}

\item{parameters}{list of parameters}

\item{covariates}{list of covariates}

\item{...}{arguments passed on to \code{PKPDsim::sim}}
}
\value{
a numeric value indicating the time, in hours, if a dosing interval
contained the target metric within range. Returns Inf if target was never
achieved. Returns NA if the target type is not supported.
}
\description{
For AUC targets, this function returns the first time of the start of
the dosing interval plus the infusion length where the AUC24 or 12 is
on target.
}
\details{
For trough targets, this returns the time of start of the dosing interval
plus the infusion length immediately prior to the trough being on target
(i.e. the dose + infusion length of the first dose that brought the trough
to target).
}
