% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sampling_scheme.R
\name{check_offset_base}
\alias{check_offset_base}
\title{Check / clean offset_base element}
\usage{
check_offset_base(offset_base, time, anchor)
}
\description{
Check / clean offset_base element
}
