% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_trial_design.R
\name{check_trial_design}
\alias{check_trial_design}
\title{Check design}
\usage{
check_trial_design(design)
}
\arguments{
\item{design}{a trial design object}
}
\value{
a trial design object
}
\description{
Check design
}
