% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_design.R
\name{create_design}
\alias{create_design}
\title{Create timing designs (static or adaptive) for use in simulated trial, such
as sampling designs, target designs etc.}
\usage{
create_design(
  time = NULL,
  when = NULL,
  offset = NULL,
  anchor = NULL,
  anchor_by = c("day", "dose")
)
}
\arguments{
\item{time}{a vector of numeric values. If no \code{anchor} values are specified,
these will be used as the fixed sampling times in the simulated trial.
If \code{anchor} values are supplied, the sampling times  will be calculated
adaptively during the trial. The \code{anchor} determine which dose is
used as reference anchor, and \code{time} will be relative to the specified
\code{anchor}.}

\item{when}{character vector of same length as \code{time} (or single
value) determining how to interpret the provided sampling \code{time}. If \code{NULL}
will use the dose time as offset (default). Other options are \code{cmax} or
\code{peak}, which will use the end of infusion as the base for the \code{time}, or
\code{cmin} or \code{trough}, which will use the time of next dose as the offset.}

\item{offset}{offset from standardized PK moments specified in \code{when}.}

\item{anchor}{numeric vector of the dose or day number to "anchor"
the sampling times to. Vector needs to be of same length as \code{time}.
If \code{anchor_by} is set to \code{day}, then the first dose in that day is used.
If later doses in the day are preferred, the anchor can also be specified
fractionally, e.g. \code{1.5} will use the time of the first dose in the
second half of the 1st day.}

\item{anchor_by}{either \code{day} or \code{dose}. Single value required, i.e. anchor
types cannot be mixed.}
}
\description{
This function is used by the three functions \code{create_sampling_design()},
\code{create_regimen_update_design()} and \code{create_target_design()}, but is itself
not intended for direct usage.
}
