% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_target_design.R
\name{create_eval_design}
\alias{create_eval_design}
\title{Create evaluation object}
\usage{
create_eval_design(
  evaltype = mipd_target_types(),
  time = NULL,
  when = NULL,
  offset = NULL,
  at = NULL,
  anchor = c("dose", "day")
)
}
\arguments{
\item{time}{a vector of numeric values at which to measure and optimize the
target. In most cases \code{time} is not required as argument, and will be inferred
from the the \code{targettype}.
If \code{at} values are supplied, the target times will be calculated
adaptively during the trial. The \code{at} determine which dose is
used as reference anchor. and \code{time} will be relative to the specified
\code{at}. If no \code{at} values are specified, the \code{time} values will be
used as the fixed absolute target times in the simulated trial.}

\item{when}{character vector of same length as \code{time} (or single
value) determining how to interpret the provided target \code{time}. If \code{NULL}
will use the dose time as offset (default). Other options are \code{cmax} or
\code{peak}, which will use the end of infusion as the base for the \code{time}, or
\code{cmin} or \code{trough}, which will use the time of next dose as the offset.}

\item{offset}{offset from a \code{when} moment (dose, peak, or trough).}

\item{at}{numeric vector of the dose or day number to "anchor"
the target times to. Vector needs to be of same length as \code{t}.
If \code{anchor} is set to \code{day}, then the first dose in that day is used.
If later doses in the day are preferred, the anchor can also be specified
fractionally, e.g. \code{1.5} will use the time of the first dose in the
second half of the 1st day.}

\item{anchor}{either \code{day} or \code{dose}. Single value required, i.e. anchor
types cannot be mixed.}
}
\description{
This function defines the evaluation metric and timing for non-target metrics.
Use this function to record outputs from the simulation like troughs or AUC
that are not at the target times.
}
