% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_initial_regimen_design.R
\name{create_initial_regimen_design}
\alias{create_initial_regimen_design}
\title{Creates a design for the initial regimen for patients in the trial}
\usage{
create_initial_regimen_design(method, ...)
}
\arguments{
\item{method}{a function to call. The function should always return a
PKPDsim regimen object.}
}
\value{
a list object with the design specs
}
\description{
Creates a design for the initial regimen for patients in the trial
}
