% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_regimen_update_design.R
\name{create_regimen_update_design}
\alias{create_regimen_update_design}
\title{Create scheme for updating dose or interval during dose optimization trial}
\usage{
create_regimen_update_design(
  time = NULL,
  anchor,
  anchor_by,
  update_type = c("dose", "interval"),
  dose_optimization_method = map_adjust_dose
)
}
\arguments{
\item{time}{a vector of numeric values. If no \code{anchor} values are specified,
these will be used as the fixed sampling times in the simulated trial.
If \code{anchor} values are supplied, the sampling times  will be calculated
adaptively during the trial. The \code{anchor} determine which dose is
used as reference anchor, and \code{time} will be relative to the specified
\code{anchor}.}

\item{anchor}{numeric vector of the dose or day number to "anchor"
the sampling times to. Vector needs to be of same length as \code{t}.
If \code{anchor_by} is set to \code{day}, then the first dose in that day is used.
If later doses in the day are preferred, the anchor can also be specified
fractionally, e.g. \code{1.5} will use the time of the first dose in the
second half of the 1st day.}

\item{anchor_by}{either \code{day} or \code{dose}. Single value required, i.e. anchor
types cannot be mixed.}

\item{update_type}{update either future \code{dose} (default) or \code{interval}.}

\item{dose_optimization_method}{A function that determines how doses should
be adjusted given the collected drug levels. The function must return a
named list of the structure \code{list(regimen = reg, additional_info = x)},
where \code{reg} is the updated PKPDsim regimen for the patient and \code{x} can be
another other information useful for post-processing of trial results.}
}
\description{
Create scheme for updating dose or interval during dose optimization trial
}
