% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_regimen_update_scheme.R
\name{create_regimen_update_scheme}
\alias{create_regimen_update_scheme}
\title{Create scheme for updating dose or interval during dose optimization trial}
\usage{
create_regimen_update_scheme(time = NULL, anchor, anchor_by)
}
\arguments{
\item{time}{a vector of numeric values. If no \code{anchor} values are specified,
these will be used as the fixed sampling times in the simulated trial.
If \code{anchor} values are supplied, the sampling times  will be calculated
adaptively during the trial. The \code{anchor} determine which dose is
used as reference anchor, and \code{time} will be relative to the specified
\code{anchor}.}

\item{anchor}{numeric vector of the dose or day number to "anchor"
the sampling times to. Vector needs to be of same length as \code{t}.
If \code{anchor_by} is set to \code{day}, then the first dose in that day is used.
If later doses in the day are preferred, the anchor can also be specified
fractionally, e.g. \code{1.5} will use the time of the first dose in the
second half of the 1st day.}

\item{anchor_by}{either \code{day} or \code{dose}. Single value required, i.e. anchor
types cannot be mixed.}
}
\description{
Create scheme for updating dose or interval during dose optimization trial
}
