% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sampling_design.R
\name{create_sampling_design}
\alias{create_sampling_design}
\title{Function for creating sampling designs.}
\usage{
create_sampling_design(
  time,
  offset_from = NULL,
  anchor = NULL,
  anchor_by = c("day", "dose")
)
}
\arguments{
\item{time}{a vector of numeric values. If no \code{anchor} values are specified,
these will be used as the fixed sampling times in the simulated trial.
If \code{anchor} values are supplied, the sampling times  will be calculated
adaptively during the trial. The \code{anchor} determine which dose is
used as reference anchor, and \code{time} will be relative to the specified
\code{anchor}.}

\item{offset_from}{character vector of same length as \code{time} (or single
value) determining how to interpret the provided sampling \code{time}. If \code{NULL}
will use the dose time as offset (default). Other options are \code{cmax} or
\code{peak}, which will use the end of infusion as the base for the \code{time}, or
\code{cmin} or \code{trough}, which will use the time of next dose as the offset.}

\item{anchor}{numeric vector of the dose or day number to "anchor"
the sampling times to. Vector needs to be of same length as \code{time}.
If \code{anchor_by} is set to \code{day}, then the first dose in that day is used.
If later doses in the day are preferred, the anchor can also be specified
fractionally, e.g. \code{1.5} will use the time of the first dose in the
second half of the 1st day.}

\item{anchor_by}{either \code{day} or \code{dose}. Single value required, i.e. anchor
types cannot be mixed.}
}
\description{
Sampling times can usually be pre-specified if the dosing schedule is fixed.
In that case a scheme can be specified with static timepoints e.g.
\code{time = c(1, 7.5, 25, 31.5)} for taking peak-trough samples at dose
1 and 4 for an 8-hour regimen.
However, when the dosing interval is adaptive, the peak and trough sampling
times cannot stay static and have to be adapted on-the-fly during
simulations.
}
\examples{
create_sampling_design(c(1.5, 11.5, 23.5))
create_sampling_design(
  time = c(0.5, -0.5),
  offset_from = c("peak", "trough"),
  anchor = c(1, 1)
)
create_sampling_design(
  time = c(0.5, -0.5, 0.5, -0.5),
  offset_from = c("peak", "trough"),
  anchor = c(1, 1, 3, 3),
  anchor_by = "dose"
)

}
