% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sampling_scheme.R
\name{create_sampling_scheme}
\alias{create_sampling_scheme}
\title{Sampling times can usually be pre-specified if the dosing schedule is fixed.
In that case a scheme can be prespecified with static timepoints e.g.
\code{c(1, 7.5, 25, 31.5)} for taking peak-trough samples at dose 1 and 4 for an
8-hour regimen.}
\usage{
create_sampling_scheme(
  time = NULL,
  anchors = NULL,
  anchor_by = c("day", "dose")
)
}
\arguments{
\item{time}{a vector of either all numeric values or all character values. If
numeric values are specified, these will be used as the fixed sampling times
in the simulated trial. If character values are supplied, the sampling times
will be calculated adaptively during the trial, if needed. The "times" should
then be of the form "\if{html}{\out{<anchor>}}<+/->\if{html}{\out{<offset>}}", e.g. "trough-0.5" or "peak+0.5",
meaning the trough for the referenced dose minus half an hour, or the peak for the
referenced dose plus half an hour. The \code{anchors} determine which dose is
used as reference anchor.}

\item{anchors}{numeric vector of the dose number of day number to "anchor"
the TDMs to. Vector needs to be of same length as \code{t}. If \code{anchor_by} is set
to \code{day}, then the first dose in that day is used. If later doses in the day
are preferred, the anchor can also be specified fractionally, e.g. \code{1.5}
will use the time of the first dose in the second half of the 1st day.}

\item{anchor_by}{either "day" or "dose". Single value required, i.e. anchor
types cannot be mixed.}
}
\description{
However, when the dosing interval is adaptive, the peak and trough sampling
times cannot stay static and have to be adapted on-the-fly during
simulations.
}
