% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_target_design.R
\name{create_target_design}
\alias{create_target_design}
\title{Create target object}
\usage{
create_target_design(
  targettype = mipd_target_types(),
  targetmin = NULL,
  targetmax = NULL,
  targetvalue = NULL,
  single_point_variation = 0.2,
  time,
  offset_from = rep("dose", length(time)),
  anchor = NULL,
  anchor_by = c("day", "dose")
)
}
\arguments{
\item{targettype}{target type, one of accepted types (see
\code{\link[=mipd_target_types]{mipd_target_types()}})}

\item{targetmin}{minimum value acceptable, must be specified with \code{targetmax}}

\item{targetmax}{maximum value acceptable, must be specified with \code{targetmin}}

\item{targetvalue}{single value for a target, overrides min and max values.}

\item{single_point_variation}{acceptable variation from targetvalue. By
default 20\%. Considered for assessment of target attainment a posteriori,
not used for dose-finding logic.}

\item{time}{a vector of numeric values at which to measure and optimize the
target. If no \code{anchor} values are specified, these will be used as the
fixed target times in the simulated trial.
If \code{anchor} values are supplied, the target times  will be calculated
adaptively during the trial. The \code{anchor} determine which dose is
used as reference anchor, and \code{time} will be relative to the specified
\code{anchor}.}

\item{offset_from}{character vector of same length as \code{time} (or single
value) determining how to interpret the provided target \code{time}. If \code{NULL}
will use the dose time as offset (default). Other options are \code{cmax} or
\code{peak}, which will use the end of infusion as the base for the \code{time}, or
\code{cmin} or \code{trough}, which will use the time of next dose as the offset.}

\item{anchor}{numeric vector of the dose or day number to "anchor"
the target times to. Vector needs to be of same length as \code{t}.
If \code{anchor_by} is set to \code{day}, then the first dose in that day is used.
If later doses in the day are preferred, the anchor can also be specified
fractionally, e.g. \code{1.5} will use the time of the first dose in the
second half of the 1st day.}

\item{anchor_by}{either \code{day} or \code{dose}. Single value required, i.e. anchor
types cannot be mixed.}
}
\description{
This function helps the user define the PKPD target for a simulation. When a
minimum value and maximum value are supplied, the algorithm targets the mid-
point. Alternatively, a single midpoint can be supplied.
}
\examples{

## Target cumulative AUC at 72 hours:
create_target_design(
  targettype = "cum_auc", targetvalue = 90,
  time = 72
)

## Target trough concentration at trough after dose 4.
create_target_design(
  targettype = "cmin",
  targetvalue = 15,
  time = 0,
  offset_from = "trough",
  anchor = 4,
  anchor_by = "dose"
)

}
