% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trial_design.R
\name{create_trial_design}
\alias{create_trial_design}
\title{Combine all sub-designs into the overall trial design object}
\usage{
create_trial_design(
  file = NULL,
  sampling_design = NULL,
  target_design = NULL,
  initial_regimen_design = NULL,
  regimen_update_design = NULL,
  sim_design = NULL,
  est_design = NULL
)
}
\arguments{
\item{file}{a YAML file with all subdesigns}

\item{sampling_design}{Design for sampling, from \code{create_sampling_design()}}

\item{target_design}{Design for target attainment, from
\code{create_target_design()}}

\item{initial_regimen_design}{Design for initial regimen, from
\code{create_initial_regimen_design()}}

\item{regimen_update_design}{Design for dose/regimen optimization, from
\code{create_regimen_update_design()}}

\item{sim_design}{Design for simulation model, from \code{create_model_design()}}

\item{est_design}{Design for estimation, from \code{create_model_design()}}
}
\description{
Combine all sub-designs into the overall trial design object
}
