% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_and_adjust.R
\name{dose_adjust_map}
\alias{dose_adjust_map}
\title{Adjust doses to achieve a target metric using MAP Bayesian estimation.}
\usage{
dose_adjust_map(
  tdms,
  est_model,
  parameters,
  omega,
  ruv,
  regimen,
  covariates,
  t_obs,
  target,
  dose_update,
  dose_grid,
  ...
)
}
\arguments{
\item{tdms}{observations to be included in MAP estimates}

\item{est_model}{model used for estimation}

\item{parameters}{population parameters of estimation model}

\item{omega}{omega matrix of estimation model}

\item{ruv}{residual error distribution of estimation model, a named list with
elements for \code{prop} (proportional) and \code{add} (additive).}

\item{regimen}{PKPDsim regimen object}

\item{covariates}{named list of PKPDsim covariates}

\item{t_obs}{vector of observation times}

\item{target}{list with arguments \code{type}, and \code{value}. \code{type} = observation
type, either "conc" or auc".}

\item{dose_update}{update dose from which dose?}

\item{dose_grid}{vector specifying doses to use as test grid, Example:
seq(from = 50, to = 500, by = (500 - 50) / 10 )}

\item{...}{arguments passed on to PKPDmap::get_map_estimates and/or
PKPDsim::sim}
}
\description{
Given a set of levels and a model definition, performs MAP Bayesian
estimation of individual PK/PD parameters, then finds the appropriate dose
to achieve the specified PK/PD target and updates the individual's regimen
accordingly.
}
