% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_grid_search.R
\name{dose_grid_search}
\alias{dose_grid_search}
\title{Perform a grid search for a particular target by simulating a grid of doses}
\usage{
dose_grid_search(
  est_model = NULL,
  regimen,
  target_time = 24,
  target = list(type = "conc", value = 10),
  auc_comp = NULL,
  pta = NULL,
  omega = NULL,
  ruv = NULL,
  dose_update = 1,
  grid = seq(1, 6000, by = 10),
  grid_type = "dose",
  dose_resolution = 1,
  refine = NULL,
  refine_range = c(0.7, 1.4),
  check_boundaries = TRUE,
  max_dose = NULL,
  min_dose = NULL,
  n_cores = 1,
  md = list(),
  covariates = NULL,
  ...
)
}
\arguments{
\item{est_model}{model used for estimation ("clinician facing")}

\item{regimen}{PKPDsim regimen object}

\item{target_time}{vector of observation times at which to evaluate the
target exposure.}

\item{target}{list with arguments \code{type}, and \code{value}. \code{type} = observation
type, either "conc" or auc".}

\item{auc_comp}{auc compartment (starting from 1, R-style not C-style!)}

\item{pta}{probability of target attainment, list with arguments \code{type} and
\code{value}, also requires \code{omega} if non-NULL. If \code{NULL}, will just aim for
specific conc or auc.}

\item{omega}{IIV matrix, for estimation model, for probability of target
attainment target types.}

\item{ruv}{list specifying residual error for estimation model:
\code{list(prop = 0.1, add = 1.5)}, for probability of target attainment target
types.}

\item{dose_update}{update dose from which dose?}

\item{grid}{vector specifying doses or intervals to use as test grid,
Example: \code{seq(from = 50, to = 500, by = (500 - 50) / 10)}}

\item{dose_resolution}{to which precision should the output be rounded
(e.g. 50), useful when in practice only a specific set of dose units.
Can of course also be controlled by altering the grid.}

\item{refine}{should the found optimal dose be refined more? If not
specified, will refine if the model linearity (in \code{attr(model, "misc")}) is
not described as \code{"linear"}}

\item{refine_range}{after initial optimization, should a second refinement
step be implemented? If \code{refine_range} is specified e.g. as \code{c(0.9, 1.1)}
then it will implement a second optimization using a grid spanning from 90\%
to 110\% of the initial optimal dose. Useful only for non-linear models.}

\item{check_boundaries}{if optimal dose is at lower/upper boundary of grid,
should grid be expanded?}

\item{max_dose}{maximum dose cap}

\item{min_dose}{minimum dose cap}

\item{n_cores}{Number of cores over which to simulate doses}

\item{md}{metadata object (only needed if we have to use
\code{get_quantity_from_variable()} to generate target value)}

\item{covariates}{covariates object}

\item{...}{passed on to PKPDsim function}
}
\value{
A numeric value indicating the recommended dose
}
\description{
Set refine = TRUE if the model is nonlinear so that the grid search happens
iteratively.
}
