% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_rows_0_100}
\alias{filter_rows_0_100}
\title{Filter rows with values 0 or 100}
\usage{
filter_rows_0_100(tab)
}
\arguments{
\item{tab}{data frame with columns \code{dose} and \code{y}}
}
\description{
Remove all rows of 0 or 1 if there are at least 2 non-(0,1) rows, and
otherwise keep all but the highest 0 row and the lowest 1 row.
}
\details{
Taken from a proprietary package, written by Kara Woo.
}
