% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_variability_terms.R
\name{generate_variability}
\alias{generate_variability}
\alias{generate_iiv}
\alias{generate_ruv}
\title{Generate variability terms}
\usage{
generate_iiv(
  sim_model,
  omega,
  parameters,
  ids = 1,
  n_iter = 1,
  seed = NULL,
  ...
)

generate_ruv(tdm_sample_time, prop, add, ids = 1, n_iter = 1, seed = NULL)
}
\arguments{
\item{sim_model}{model used for simulated patient response ("truth").}

\item{omega}{omega matrix, with covariance terms.
See \code{PKPDsim::sim} for details.}

\item{parameters}{simulation model parameters (population estimates), a
named list.}

\item{ids}{vector of ids, can be numeric or character.}

\item{n_iter}{number of sets of individual parameters to generate per id}

\item{seed}{set random seed}

\item{...}{arguments passed on to PKPDsim::sim}

\item{tdm_sample_time}{time of tdm, since start of treatment course (or
other vector of identifiers to use for each tdm). For example, for three
days of daily dosing and peak-trough sample collection,
\code{c(1, 24, 25, 48, 49, 73)}.}

\item{prop}{proportional error}

\item{add}{additive error}
}
\value{
\code{generate_iiv} a data frame with columns \code{id} (corresponding to
\code{ids}), \code{iter} ( numbers 1 to n_iter) and columns for each individual
parameter value.

\code{generate_ruv} returns a data frame with identifier columns of
\code{tdm_number}, \code{iteration}, \code{id}, plus columns for proportional (\code{prop}) and
additive (\code{add}) error.
}
\description{
To enable comparison of multiple treatment conditions in a reproducible
manner, it is recommended that interindividual variability terms and residual
variability terms be generated prior to all analyses. This design also allows
for resuming a simulation part-way through, when the random seed position
may not be known.

Generate IIV for one or more individuals and one or more iterations per
individual according to the supplied omega matrix.

Generate unexplained variability for one or more individuals and one or
more iterations per individual according to the supplied proportional and
additive error.
}
\details{
This family of functions generates variability terms to allow for
reproducible analyses. Using multiple iterations per individual ID allows for
PK variability within one set of covariates.

By default, \code{generate_iiv}assumes a log-normal (exponential) distribution.
See \code{PKPDsim::sim} documentation for the \code{omega_type} argument to provide
finer grain control.

\code{generate_ruv} Assumes a normal distribution for proportional and
additional error.
}
