% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dose_update_numbers_from_scheme.R
\name{get_dose_update_core}
\alias{get_dose_update_core}
\title{Core function to calculate the dose update number for a row in a
regimen update data.frame}
\usage{
get_dose_update_core(row, regimen)
}
\arguments{
\item{row}{a single row from a regimen_update data.frame}

\item{regimen}{PKPDsim regimen}
}
\description{
Core function to calculate the dose update number for a row in a
regimen update data.frame
}
