% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quantity_from_variable.R
\name{get_quantity_from_variable}
\alias{get_quantity_from_variable}
\title{Get quantities from variables in sim results}
\usage{
get_quantity_from_variable(var, sim, md, times, comp = NULL)
}
\arguments{
\item{var}{Name of variable}

\item{sim}{Output of PKPDsim::sim_ode()}

\item{md}{Metadata object}

\item{times}{Times at which to calculate}

\item{comp}{if not NULL, filter by \code{sim$comp}}
}
\description{
Used to add quantities to rows in advice table, exposure time table.
}
