% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sampling_times_from_schema.R
\name{get_sampling_time_core}
\alias{get_sampling_time_core}
\title{Core function to calculate the sampling time for a row in a sampling schema
data.frame.}
\usage{
get_sampling_time_core(row, regimen)
}
\arguments{
\item{row}{a single row from a sampling_time_design data.frame}

\item{regimen}{PKPDsim regimen}
}
\description{
Core function to calculate the sampling time for a row in a sampling schema
data.frame.
}
