% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sampling_times_from_schema.R
\name{get_sampling_times_from_scheme}
\alias{get_sampling_times_from_scheme}
\title{Calculate sampling times based on a given sampling schema and a regimen.}
\usage{
get_sampling_times_from_scheme(scheme, regimen)
}
\arguments{
\item{scheme}{sampling schema created using \code{create_tdm_schema()}}

\item{regimen}{a \code{PKPDsim} regimen object}
}
\value{
vector of numeric sampling times
}
\description{
Calculate sampling times based on a given sampling schema and a regimen.
}
