% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_single_target_design.R
\name{get_single_target_design}
\alias{get_single_target_design}
\title{Get a single target from a potentially time-varying target design}
\usage{
get_single_target_design(target_design, idx = nrow(target_design$scheme))
}
\arguments{
\item{target_design}{target design, created using \code{create_target_design}.}

\item{idx}{number of regimen update. By default it will take the last
instance.}
}
\description{
In most cases, the target design will just specify a single target.
But in some scenarios, a time-varying target may make sense. In that
case, for each regimen update instance, a different target can be
specified.
This function just grabs the right target values from the overall
target design. If just a single target is specified, it will just
return the target design as-is.
}
