% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_target_design.R
\name{is_on_target}
\alias{is_on_target}
\title{Checks if a value (or vector of values) is within the specified target range}
\usage{
is_on_target(v, target)
}
\arguments{
\item{v}{exposure metric, a single value or a vector}

\item{target}{target specification created with \code{\link[=create_target_design]{create_target_design()}}, or
a named list with \code{min} and \code{max} specified.}
}
\value{
Returns a logical value of \code{TRUE} or \code{FALSE} for each value in v.
}
\description{
Checks if a value (or vector of values) is within the specified target range
}
