% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_number}
\alias{is_valid_number}
\title{Checks that an object represents a single finite number}
\usage{
is_valid_number(n)
}
\arguments{
\item{n}{an object to check}
}
\description{
Checks that an object represents a single finite number
}
\examples{
mipdtrial:::is_single_valid_number(9)
mipdtrial:::is_single_valid_number(mtcars)
mipdtrial:::is_single_valid_number(c(1, 2))
}
