% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_numeric_vector}
\alias{is_valid_numeric_vector}
\title{Checks that an object represents a vectir of finite number
with no NA or NaN or Inf}
\usage{
is_valid_numeric_vector(x)
}
\arguments{
\item{n}{an object to check}
}
\description{
Checks that an object represents a vectir of finite number
with no NA or NaN or Inf
}
\examples{
mipdtrial:::is_valid_numeric_vector(9)
mipdtrial:::is_valid_numeric_vector(mtcars)
mipdtrial:::is_valid_numeric_vector(c(1, 2))
}
