% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_target_object.R
\name{mipd_target_types}
\alias{mipd_target_types}
\title{Accepted PK/PD exposure targets}
\usage{
mipd_target_types()
}
\value{
Returns a character vector of accepted target types.
}
\description{
Model-based dose-finding is currently implemented for the following target
types:
\itemize{
\item cmax: Peak concentration
\item cmax_1hr: Peak concentration 1hr after dose
\item ctrough, cmin: Trough concentration
\item conc: generic concentration
\item cum_auc: Cumulative AUC
\item auc: auc over a dosing interval
\item auc24: auc normalized to a 24-hour period
}
}
