% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipdtrial-package.R
\docType{package}
\name{mipdtrial-package}
\alias{mipdtrial}
\alias{mipdtrial-package}
\title{MIPDtrial package}
\description{
Simulation model-informed precision dosing trials
}
\author{
Jasmine Hughes \email{jasmine@insight-rx.com}
}
