% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_based_starting_dose.R
\name{model_based_starting_dose}
\alias{model_based_starting_dose}
\title{Model-based starting dose}
\usage{
model_based_starting_dose(design, covariates, cov_mapping, ...)
}
\arguments{
\item{design}{trial design, created using \code{create_trial_design()}}

\item{covariates}{named list of PKPDsim covariates. E.g.:
list(WT = PKPDsim::new_covariate(70))}

\item{cov_mapping}{a named vector indicating correspondence between model
covariates and data frame columns. Vector names should reflect all
covariates defined by the model, and vector values should represent the
associated column in the provided data set.}

\item{...}{arguments passed on to \code{dose_grid_search}}
}
\value{
Returns a PKPDsim regimen populated with the model-predicted dose
}
\description{
A light wrapper for \link{dose_grid_search} for finding model-based starting doses
and returning an appropriate regimen object.
}
