% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trial_design.R
\name{parse_spec_file_to_trial_design}
\alias{parse_spec_file_to_trial_design}
\title{Parse YAML spec file to trial design}
\usage{
parse_spec_file_to_trial_design(file)
}
\arguments{
\item{file}{full filepath to yaml file}
}
\description{
Parse YAML spec file to trial design
}
