% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_nca.R
\name{perform_nca}
\alias{perform_nca}
\title{Perform non-compartmental analysis (NCA)}
\usage{
perform_nca(tdms, regimen, ...)
}
\arguments{
\item{tdms}{a data frame of TDMs, containing at least a column \code{t} for time
and \code{y} for observed measurement to be used in estimation. Will filter for
the samples from the final dosing interval. Baseline estimates must be
collected within 30 minutes of the time of the start of dose administration
in order to be considered baseline values.}

\item{regimen}{A PKPDsim regimen object}

\item{...}{arguments passed on to \code{clinPK::nca}.}
}
\value{
Returns an object of class "nca_output", defined in \code{clinPK}. This
object is a named list, with components for \code{pk} (containing PK
parameters), \code{descriptive} (containing exposure estimates) and \code{settings}
(containing input parameters to the NCA calculations). See \code{help("nca")}
for more information.
}
\description{
A light wrapper around \code{clinPK::nca}. Given a set of TDMs, performs
non-compartmental analysis. NCA only considers TDMs from a single dosing
interval. This function identifies the dose corresponding to the final TDM(s)
and uses all TDMs following that dose.
}
