% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_yaml_safe}
\alias{read_yaml_safe}
\title{Safe way to read YAML}
\usage{
read_yaml_safe(filename)
}
\arguments{
\item{filename}{YAML file to parse}
}
\description{
Workaround to avoid issue with "n" entries
https://github.com/vubiostat/r-yaml/issues/122)
}
