% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_to_multiple}
\alias{round_to_multiple}
\title{Round to a multiple of any number (e.g. round to the nearest 5, 10, 100)}
\usage{
round_to_multiple(x, multiple)
}
\arguments{
\item{x}{value to be rounded}

\item{multiple}{accuracy to round to. If NULL, x will be returned unrounded.}
}
\description{
Round to a multiple of any number (e.g. round to the nearest 5, 10, 100)
}
\examples{
# Uses "round-to-even" strategy
round_to_multiple(12.5, 5)
round_to_multiple(17.5, 5)
}
