% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trial.R
\name{run_trial}
\alias{run_trial}
\title{Run an MIPD trial}
\usage{
run_trial(
  data,
  design,
  cov_mapping,
  n_ids = NULL,
  seed = NULL,
  progress = TRUE
)
}
\arguments{
\item{n_ids}{number of subjects to use in simulated trial. If not specified,
will use all subjects in \code{data}.}

\item{progress}{show progress bar? (default is \code{TRUE})}
}
\description{
This function is a wrapper function to make it easier to run a basic MIPD
trial with minimal function calls. For fine-tune control over logic, it may
be easier to compose each step of the precision-dosing workflow yourself. For
an example, see the "Busulfan MIPD" vignette.
}
\details{
This function performs the following steps:
\enumerate{
\item Dose the patient using the function specified in \code{initial_dose_method}.
\item Collect levels at the specified times (\code{tdm_times}).
\item Adjust dosing at each specified dose (\code{adjust_doses_at}) using the logic
specified in \code{dose_optimization_method}
\item Return an object with data collected during and at the end of the MIPD
trial.
}
}
