% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_grid_search.R
\name{simulate_dose}
\alias{simulate_dose}
\title{Simulate dose}
\usage{
simulate_dose(
  dose_grid,
  dose_update,
  regimen,
  md,
  pta,
  target,
  model,
  t_obs,
  omega,
  obs,
  ruv,
  ...
)
}
\arguments{
\item{dose_grid}{element of the dose grid}

\item{dose_update}{update dose from which dose?}

\item{regimen}{regimen}

\item{md}{metadata object (only needed if we have to use \code{get_quantity_from_variable()} to generate target value)}

\item{pta}{probability of target attainment, list with arguments \code{type} and \code{value}, also requires \code{omega} if non-NULL. If \code{NULL}, will just aim for specific conc or auc.}

\item{target}{list with arguments \code{type}, \code{method}, and \code{value}. \code{type} = observation type, either "conc" or auc".}

\item{model}{model for simulating dose (estimation model)}

\item{t_obs}{vector of observation times}

\item{omega}{IIV matrix, for estimation model}

\item{obs}{Value of \code{obs} as determined by \code{\link[=dose_grid_search]{dose_grid_search()}} (i.e. either "obs" or AUC compartment)}

\item{ruv}{list specifying residual error for estimation model: \code{list(prop = 0.1, add = 1.5)}}

\item{...}{passed on to PKPDsim function}
}
\description{
Simulate a dose in the dose grid
}
