% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_grid_search.R
\name{simulate_dose_interval}
\alias{simulate_dose_interval}
\title{Simulate different doses/intervals in a dose/interval grid}
\usage{
simulate_dose_interval(
  value,
  grid_type = "dose",
  dose_update,
  regimen,
  md,
  pta,
  target_design,
  model,
  t_obs,
  omega,
  obs,
  ruv,
  ...
)
}
\arguments{
\item{value}{element of the dose/interval grid}

\item{grid_type}{either \code{dose} grid or \code{interval} grid}

\item{dose_update}{update dose from which dose?}

\item{regimen}{PKPDsim regimen object}

\item{md}{metadata object (only needed if we have to use
\code{get_quantity_from_variable()} to generate target value)}

\item{pta}{probability of target attainment, list with arguments \code{type} and
\code{value}, also requires \code{omega} if non-NULL. If \code{NULL}, will just aim for
specific conc or auc.}

\item{target_design}{object specifying target design, created using function
\code{create_target_design()}}

\item{model}{model for simulating dose (estimation model)}

\item{t_obs}{time at which observation should be calculated}

\item{omega}{IIV matrix, for estimation model, for probability of target
attainment target types.}

\item{obs}{Value of \code{obs} as determined by \code{\link[=dose_grid_search]{dose_grid_search()}} (i.e. either "obs" or AUC compartment)}

\item{ruv}{list specifying residual error for estimation model:
\code{list(prop = 0.1, add = 1.5)}, for probability of target attainment target
types.}

\item{...}{passed on to PKPDsim function}
}
\description{
Simulate different doses/intervals in a dose/interval grid
}
