% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_fit.R
\name{simulate_fit}
\alias{simulate_fit}
\title{Get MAP Bayesian parameters}
\usage{
simulate_fit(est_model, parameters, omega, ruv, tdms, covariates, regimen)
}
\arguments{
\item{est_model}{model used for estimation}

\item{parameters}{population parameters of estimation model}

\item{omega}{omega matrix of estimation model}

\item{ruv}{residual error distribution of estimation model, a named list with
elements for \code{prop} (proportional) and \code{add} (additive).}

\item{tdms}{observations to be included in MAP estimates}

\item{covariates}{named list of PKPDsim covariates}

\item{regimen}{PKPDsim regimen object}
}
\value{
named list of individual PK parameter estimates
}
\description{
A light wrapper around \code{PKPDmap::get_map_estimates} and
\code{PKPDmap::create_iov_object}. For more fine-grained control, use \code{PKPDmap}
functions directly.
}
\details{
MAP Bayesian fits sometimes fail when simulated values are very unlikely. To
prevent a long simulation from stalling, fit is wrapped in a tryCatch, and
NULL is returned for appropriate processing outside the function.
}
