% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_nca.R
\name{tau_from_regimen}
\alias{tau_from_regimen}
\title{Get dosing interval (tau) from PKPDsim regimen object}
\usage{
tau_from_regimen(regimen, dose_n)
}
\arguments{
\item{regimen}{A PKPDsim regimen object}

\item{dose_n}{index of dose for which tau is requested}
}
\value{
a single numeric value
}
\description{
Uses time of next dose for all but the last interval, which takes from the
PKPDsim regimen interval specification.
}
