% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_target_to_variable}
\alias{time_target_to_variable}
\title{Map time-based target type to model variable name}
\usage{
time_target_to_variable(target_type)
}
\arguments{
\item{target_type}{character, one of the time-based target types
(\code{"t_gt_mic"}, \code{"t_gt_4mic"}, \code{"t_gt_mic_free"}, \code{"t_gt_4mic_free"})}
}
\value{
A character string with the corresponding model variable name,
or \code{NULL} if the target type is not recognized.
}
\description{
Map time-based target type to model variable name
}
