% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_regimen.R
\name{update_regimen}
\alias{update_regimen}
\title{Update a regimen with a new dose}
\usage{
update_regimen(
  regimen,
  new_dose = NULL,
  new_interval = NULL,
  dose_update_number
)
}
\arguments{
\item{regimen}{PKPDsim regimen object}

\item{new_dose}{value of new dose}

\item{new_interval}{value of new interval}

\item{dose_update_number}{integer indicating which dose and onwards should
be updated}
}
\value{
Returns a PKPDsim regimen object with the new dose applied.
}
\description{
Update a regimen with a new dose
}
