% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosing_nomograms.R
\name{weight_based_starting_dose}
\alias{weight_based_starting_dose}
\title{Weight-based starting dose (e.g., mg/kg)}
\usage{
weight_based_starting_dose(covariates, design, cov_mapping)
}
\arguments{
\item{args}{something}
}
\description{
Weight-based starting dose (e.g., mg/kg)
}
