% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_and_adjust.R
\name{bind_results_from_adjustments}
\alias{bind_results_from_adjustments}
\title{Bind together the results from sampling and dose adjusting}
\usage{
bind_results_from_adjustments(
  out,
  j,
  regimen,
  adjust_at_dose,
  dose_before_update,
  auc_final,
  trough_final,
  tgt_final
)
}
\arguments{
\item{out}{output object}

\item{j}{index number}

\item{regimen}{regimen}

\item{adjust_at_dose}{adjust at dose number}

\item{dose_before_update}{dose before update}

\item{auc_final}{final AUC}

\item{trough_final}{final Ctrough}

\item{tgt_final}{final time target percentage}
}
\value{
a data.frame with results
}
\description{
Bind together the results from sampling and dose adjusting
}
