% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_and_adjust.R
\name{bind_results_from_adjustments}
\alias{bind_results_from_adjustments}
\title{Bind together the results from sampling and dose adjusting}
\usage{
bind_results_from_adjustments(
  out,
  dose_updates,
  j,
  regimen,
  adjust_at_dose,
  dose_before_update,
  auc_final,
  trough_final
)
}
\value{
a data.frame with results
}
\description{
Bind together the results from sampling and dose adjusting
}
