% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_metrics.R
\name{calc_auc_from_sim}
\alias{calc_auc_from_sim}
\title{Get AUC from a simulation}
\usage{
calc_auc_from_sim(sim_output, auc_comp, target_time, target_type)
}
\arguments{
\item{sim_output}{output of a \code{PKPDsim::sim} call}

\item{auc_comp}{auc compartment}

\item{target_time}{target time from \code{get_sampling_times_from_scheme}}

\item{target_type}{type of AUC target in \code{c(auc24, auc12, cum_auc)}.}
}
\value{
\code{calc_auc_from_sim} returns a numeric vector of AUCs between each
simulated time point. Control time period over which AUC should be
calculated using \code{target_time} argument to \code{PKPDsim::sim}.
}
\description{
Get AUC from a simulation
}
