% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_metrics.R
\name{calc_auc_from_sim}
\alias{calc_auc_from_sim}
\title{Get AUC from a simulation}
\usage{
calc_auc_from_sim(sim_output, auc_comp, extract_time, target_type)
}
\arguments{
\item{sim_output}{output of a \code{PKPDsim::sim} call}

\item{auc_comp}{auc compartment}

\item{extract_time}{time points for AUC12 or 24 to extract}

\item{target_type}{type of AUC target in \code{c("auc24", "auc12", "cum_auc")}.}
}
\value{
\code{calc_auc_from_sim} returns a numeric vector of AUCs between each
simulated time point. Control time period over which AUC should be
calculated using \code{target_time} argument to \code{PKPDsim::sim}.
}
\description{
Get AUC from a simulation
}
