% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_metrics.R
\name{calc_time_to_target}
\alias{calc_time_to_target}
\title{Calculate time to target attainment}
\usage{
calc_time_to_target(
  regimen,
  target_design,
  auc_comp,
  model,
  terminal_interval = 24,
  ...
)
}
\arguments{
\item{regimen}{PKPDsim regimen object}

\item{target_design}{target design, created using \code{create_target_design()}}

\item{auc_comp}{compartment to look for AUC}

\item{model}{PKPDsim model object}

\item{terminal_interval}{duration of final dosing interval}

\item{...}{arguments passed on to \code{PKPDsim::sim}}
}
\value{
a numeric value indicating the time, in hours, if a dosing interval
contained the target metric within range. Otherwise, returns NA.
}
\description{
Calculate time to target attainment
}
