% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_tdms.R
\name{collect_tdms}
\alias{collect_tdms}
\title{Simulate TDM collection}
\usage{
collect_tdms(sim_model, t_obs, res_var, pars_i, lloq = NULL, ...)
}
\arguments{
\item{sim_model}{model used for simulated patient response ("truth").}

\item{t_obs}{sample collection times (since start of treatment)}

\item{res_var}{data frame of residual variability quantities to add. See
\code{pregenerate_ruv} for details. Expects columns \code{prop} and \code{add}. Error is
added in the linear domain, so log-transformed models should supply error
as proportional error with 0 additive error.}

\item{pars_i}{true parameters for the individual (named list)}

\item{lloq}{lower limit of quantification. If non-NULL, all TDMs below LLOQ
will be set to half the LLOQ.}

\item{...}{arguments passed on to PKPDsim::sim}
}
\value{
a data frame with columns \code{t} (time), \code{true_y} (actual level) and
\code{y} (measured level), with rows corresponding to t_obs.
}
\description{
Using the "ground truth" model, simulate collection of drug/biomarker levels.
Returns a data frame of sample time, true value, and measured value (with
residual error added).
}
