% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_design.R
\name{create_design}
\alias{create_design}
\title{Create timing designs (static or adaptive) for use in simulated trial, such
as sampling designs, target designs etc.}
\usage{
create_design(
  time = NULL,
  when = NULL,
  offset = NULL,
  scatter = NULL,
  at = NULL,
  anchor = c("dose", "day")
)
}
\arguments{
\item{time}{a vector of numeric values. If no \code{at} values are specified,
these will be used as the fixed sampling times in the simulated trial.
If \code{at} values are supplied, the sampling times  will be calculated
adaptively during the trial. The \code{at} determine which dose or day is
used as reference, and \code{time} will be relative to the specified \code{at} anchor.}

\item{when}{character vector of same length as \code{time} (or single
value) determining how to interpret the provided sampling \code{time}. If \code{NULL}
will use the dose time as offset (default). Other options are: \code{cmax} or
\code{peak}, which will use the end of infusion as the base for the \code{time}, or
\code{cmin} or \code{trough}, which will use the time of next dose as the offset, or
\code{middle} or \code{cmid} which will use the middle between the anchored dose and
the next, or \code{random} which takes a random time point between the anchored
dose and the next.}

\item{offset}{offset from standardized PK moments specified in \code{when}, e.g.
\code{c(1, -1)} with \code{when = c("peak", "trough")} to sample 1 hour after peak and
1 hour before trough.}

\item{scatter}{optional random variation in time, specified as the standard
deviation, e.g. \code{scatter = 0.1} to allow for variation in sampling time with
an SD of 0.1 hours. Only relevant for sampling times, not for
regimen_update designs or target designs. Random variation does not protect
for peaks or troughs becoming sampled during infusion or in previous / next
dose. So value for \code{scatter} should be chosen appropriately and probably
used in conjunction with approriate \code{offset} values.}

\item{at}{numeric vector of the dose or day number to "anchor"
the sampling times to. Vector needs to be of same length as \code{time}.
If \code{anchor} is set to \code{day}, then the first dose in that day is used.
If later doses in the day are preferred, the anchor can also be specified
fractionally, e.g. \code{1.5} will use the time of the first dose in the
second half of the 1st day.}

\item{anchor}{either \code{day} or \code{dose}. Single value required, i.e. anchor
types cannot be mixed.}
}
\description{
This function is used by the three functions \code{create_sampling_design()},
\code{create_regimen_update_design()} and \code{create_target_design()}, but is itself
not intended for direct usage.
}
