% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_design.R
\name{create_model_design}
\alias{create_model_design}
\title{Create a design for the model, including population and random-effects
parameters.}
\usage{
create_model_design(
  lib = NULL,
  model = NULL,
  parameters = NULL,
  omega_matrix = NULL,
  ruv = NULL
)
}
\description{
Create a design for the model, including population and random-effects
parameters.
}
