% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_target_object.R
\name{create_target_object}
\alias{create_target_object}
\title{Create target object}
\usage{
create_target_object(
  targettype = mipd_target_types(),
  targetmin = NULL,
  targetmax = NULL,
  targetvalue = NULL
)
}
\arguments{
\item{targettype}{target type, one of accepted types (see
\code{\link[=mipd_target_types]{mipd_target_types()}})}

\item{targetmin}{minimum value acceptable, must be specified with \code{targetmax}}

\item{targetmax}{maximum value acceptable, must be specified with \code{targetmin}}

\item{targetvalue}{single value for a target, overrides min and max values.}
}
\description{
This function helps the user define the PKPD target for a simulation. When a
minimum value and maximum value are supplied, the algorithm targets the mid-
point. Alternatively, a single midpoint can be supplied.
}
\examples{
create_target_object(targettype = "cum_auc", targetvalue = 90)

}
