% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_and_adjust.R
\name{dose_adjust_nca}
\alias{dose_adjust_nca}
\title{Adjust doses to achieve a target metric using NCA}
\usage{
dose_adjust_nca(tdms, regimen, target, dose_update, additional_info, ...)
}
\arguments{
\item{tdms}{observations to be included in MAP estimates}

\item{regimen}{PKPDsim regimen object}

\item{target}{list with arguments \code{type}, and \code{value}. \code{type} = observation
type, either "conc" or auc".}

\item{dose_update}{update dose from which dose?}

\item{additional_info}{object returned and iteratively built in
\code{sample_and_adjust_by_dose}. See \code{returns}}

\item{...}{arguments passed on to \code{clinPK::nca}}
}
\value{
Returns a named list: \code{regimen}: the updated regimen;
\code{additional_info}: output of clinPK::nca with a field for cumulative auc by
that dose.
}
\description{
Given a set of levels calculates the AUC for a given dose, and adjusts the
subsequent doses to achieve either the target AUC per dosing interval
(\code{create_target_object(targettype = 'auc')}), AUC24
(\code{create_target_object(targettype = 'auc24')}) or cumulative AUC
(\code{create_target_object(targettype = 'cum_auc')}). Uses the most recently
sampled dosing interval to estimate
}
