% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_and_adjust.R
\name{dose_adjust_nca}
\alias{dose_adjust_nca}
\title{Adjust doses to achieve a target metric using NCA}
\usage{
dose_adjust_nca(tdms, regimen, target_time, target, dose_update, ...)
}
\arguments{
\item{tdms}{observations to be included in MAP estimates}

\item{regimen}{PKPDsim regimen object}

\item{target_time}{vector of observation times at which to evaluate the
target exposure.}

\item{target}{list with arguments \code{type}, and \code{value}. \code{type} = observation
type, either "conc" or auc".}

\item{dose_update}{update dose from which dose?}

\item{...}{arguments passed on to \code{clinPK::nca}}
}
\value{
Returns a named list: \code{regimen}: the updated regimen;
\code{additional_info}: the MAP parameter estimates
}
\description{
Given a set of levels and a model definition, performs MAP Bayesian
estimation of individual PK/PD parameters, then finds the appropriate dose
to achieve the specified PK/PD target and updates the individual's regimen
accordingly.
}
