% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_nca.R
\name{dose_from_auc}
\alias{dose_from_auc}
\title{Calculate a new dose given an estimated AUC and target AUC}
\usage{
dose_from_auc(target, intv_auc, regimen, dose_update, cum_auc = 0)
}
\arguments{
\item{target}{object describing exposure targets, see \code{create_target_object}.}

\item{intv_auc}{AUC estimated for a sampled dosing interval}

\item{regimen}{PKPDsim regimen object}

\item{dose_update}{index of dose that should be updated}

\item{cum_auc}{AUC estimated for all dosing intervals up to the dose_update
interval.}
}
\description{
Uses a simple ratio of \code{dose2 = dose1 * auc2 / auc1} (i.e., assumes
constant clearance) to identify a new dose. For cumulative AUC targets, takes
into consideration previous AUC and number of doses remaining.
}
