% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tdm_times_from_schema.R
\name{get_sampling_time_core}
\alias{get_sampling_time_core}
\title{Core function to calculate the sampling time for a row in a sampling schema
data.frame.}
\usage{
get_sampling_time_core(row, regimen)
}
\description{
Core function to calculate the sampling time for a row in a sampling schema
data.frame.
}
