% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_and_adjust.R
\name{map_adjust_dose}
\alias{map_adjust_dose}
\title{Adjust doses to achieve a target metric using MAP Bayesian estimation.}
\usage{
map_adjust_dose(
  tdms,
  est_model,
  parameters,
  omega,
  ruv,
  regimen,
  covariates = NULL,
  target_design,
  dose_update,
  grid = NULL,
  ...
)
}
\arguments{
\item{tdms}{observations to be included in MAP estimates}

\item{est_model}{model used for estimation}

\item{parameters}{population parameters of estimation model}

\item{omega}{omega matrix of estimation model}

\item{ruv}{residual error distribution of estimation model, a named list with
elements for \code{prop} (proportional) and \code{add} (additive).}

\item{regimen}{PKPDsim regimen object}

\item{covariates}{named list of PKPDsim covariates}

\item{target_design}{object specifying target design, created using function
\code{create_target_design()}}

\item{dose_update}{update dose from which dose?}

\item{grid}{vector specifying doses or intervals to use as test grid,
Example: \code{seq(from = 50, to = 500, by = (500 - 50) / 10)}}

\item{...}{arguments passed on to PKPDmap::get_map_estimates and/or
PKPDsim::sim}
}
\value{
Returns a named list: \code{regimen}: the updated regimen;
\code{est_parameters}: the MAP parameter estimates
}
\description{
Given a set of levels and a model definition, performs MAP Bayesian
estimation of individual PK/PD parameters, then finds the appropriate dose
to achieve the specified PK/PD target and updates the individual's regimen
accordingly.
}
