% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_and_adjust.R
\name{map_adjust_interval}
\alias{map_adjust_interval}
\title{Adjust intervals to achieve a target metric using MAP Bayesian estimation by
adapting the dosing interval}
\usage{
map_adjust_interval(
  tdms,
  est_model,
  parameters,
  omega,
  ruv,
  regimen,
  covariates = NULL,
  target_time,
  target,
  dose_update,
  interval_grid = NULL,
  ...
)
}
\arguments{
\item{tdms}{observations to be included in MAP estimates}

\item{est_model}{model used for estimation}

\item{parameters}{population parameters of estimation model}

\item{omega}{omega matrix of estimation model}

\item{ruv}{residual error distribution of estimation model, a named list with
elements for \code{prop} (proportional) and \code{add} (additive).}

\item{regimen}{PKPDsim regimen object}

\item{covariates}{named list of PKPDsim covariates}

\item{target}{list with arguments, created using function
\code{create_target_object()}}

\item{dose_update}{update dose from which dose?}

\item{...}{arguments passed on to PKPDmap::get_map_estimates and/or
PKPDsim::sim}
}
\value{
Returns a named list: \code{regimen}: the updated regimen;
\code{additional_info}: the MAP parameter estimates
}
\description{
Given a set of levels and a model definition, performs MAP Bayesian
estimation of individual PK/PD parameters, then finds the appropriate dosing
interval to achieve the specified PK/PD target and updates the
individual's regimen accordingly.
}
