% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_based_starting_dose.R
\name{model_based_starting_dose}
\alias{model_based_starting_dose}
\title{Model-based starting dose}
\usage{
model_based_starting_dose(design, covariates, cov_mapping, ...)
}
\arguments{
\item{...}{arguments passed on to \code{dose_grid_search}}
}
\value{
Returns a PKPDsim regimen populated with the model-predicted dose
}
\description{
A light wrapper for \link{dose_grid_search} for finding model-based starting doses
and returning an appropriate regimen object.
}
