% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_grid_search.R
\name{model_based_starting_dose}
\alias{model_based_starting_dose}
\title{Model-based starting dose}
\usage{
model_based_starting_dose(n, interval, type = "infusion", t_inf = 1, ...)
}
\arguments{
\item{n}{number of doses}

\item{interval}{dosing interval}

\item{type}{dose type, e.g.: "infusion", "oral".
(see \code{\link[PKPDsim:new_regimen]{PKPDsim::new_regimen()}})}

\item{t_inf}{infusion time
... arguments passed on to \code{dose_grid_search}}
}
\value{
Returns a PKPDsim regimen populated with the model-predicted dose
}
\description{
A light wrapper for \link{dose_grid_search} for finding model-based starting doses
and returning an appropriate regimen object.
}
