% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trial.R
\name{run_trial}
\alias{run_trial}
\title{Run an MIPD trial}
\usage{
run_trial(
  data,
  design,
  cov_mapping,
  n_ids = NULL,
  seed = 0,
  verbose = FALSE,
  progress = TRUE,
  threads = 1
)
}
\arguments{
\item{data}{data set to use for simulation, a data frame. Should contain
covariates in columns, corresponding to values in \code{cov_mapping} and a
column for \code{ID}, as a patient identifier.}

\item{design}{trial design, created using \code{create_trial_design()}}

\item{cov_mapping}{a named vector indicating correspondence between model
covariates and data frame columns. Vector names should reflect all
covariates defined by the model, and vector values should represent the
associated column in the provided data set.}

\item{n_ids}{number of subjects to use in simulated trial. If not specified,
will use all subjects in \code{data}.}

\item{seed}{argument passed to \code{set.seed()} to set the seed for all random
variability generation to ensure reproducibility of results.
Even when run in parallel, reproducibility is ensures since each
thread will receive a reproducible seed number.}

\item{verbose}{verbose output?}

\item{progress}{show progress bar? (default is \code{TRUE})}

\item{threads}{number of threads to run the simulations on. By default the
simulations will run on 4 cores, or less if less are available (one core will
always be reserved). The user can override by specifying manually.}
}
\description{
This function is a wrapper function to make it easier to run a basic MIPD
trial with minimal function calls. For fine-tune control over logic, it may
be easier to compose each step of the precision-dosing workflow yourself. For
an example, see the "Busulfan MIPD" vignette.
}
\details{
This function performs the following steps:
\enumerate{
\item Dose the patient using the function specified in \code{initial_dose_method}.
\item Collect levels at the specified times (\code{tdm_times}).
\item Adjust dosing at each specified dose (\code{adjust_doses_at}) using the logic
specified in \code{dose_optimization_method}
\item Return an object with data collected during and at the end of the MIPD
trial.
}
}
