% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_and_adjust.R
\name{sample_and_adjust_by_dose}
\alias{sample_and_adjust_by_dose}
\title{Adjust dosing using MIPD on TDMs at specified dose numbers}
\usage{
sample_and_adjust_by_dose(
  regimen_update_design,
  sampling_design,
  target_design,
  regimen,
  covariates = NULL,
  pars_true_i,
  sim_model,
  sim_ruv = NULL,
  verbose = FALSE,
  accumulate_data = TRUE,
  est_design = NULL,
  ...
)
}
\arguments{
\item{regimen_update_design}{a data.frame with scheme with details on how
and when to update the regimen in the MIPD trial. Schemes are created using
\code{create_regimen_update_design()}.}

\item{sampling_design}{a data.frame with a sampling scheme, created using
\code{create_sampling_design()}.}

\item{target_design}{target design created using \code{create_target_design()}}

\item{regimen}{PKPDsim regimen object, containing initial dosing regimen.}

\item{covariates}{named list of PKPDsim covariates.}

\item{pars_true_i}{PK parameters for the individual. See \code{generate_iiv}.}

\item{sim_model}{model to use for simulating "true" patient response.}

\item{sim_ruv}{residual variability for \code{sim_model}. Named list for
proportional (\code{prop}) and additive (\code{add}) error.}

\item{verbose}{verbose output?}

\item{accumulate_data}{if \code{TRUE}, will use all available data up until the
adjustment timepoint. If set to \code{FALSE}, will use only the data since the
last adjustment timepoint and the current one.}

\item{est_design}{design specs for model used in MAP estimation (optional).}

\item{...}{arguments passed on to \code{simulate_fit} or dose_optimization_method
function.}
}
\value{
a named list containing \code{final_regimen} (all doses after
adjustment), \code{tdms} (all collected levels, both true and measured, that is,
both with and without residual variability), and \code{additional_info}, which
varies by dose_optimization_method. See selected function for details.
}
\description{
Doses are iteratively adjusted for the specified dose numbers. Estimation is
performed by MAP Bayesian estimation, and simulation is used to identify the
dose most likely to reach the specified target. See \code{dose_grid_search} for
dose-finding logic. Optionally, disparate models can be used for simulation
and estimation, to allow for intentional model misspecification. Covariates
supplied must support both models. True patient parameters (\code{pars_true_i})
must match the model used for simulation.
}
\details{
Trial design or MIPD protocol can vary considerably in complexity. This
function is intended to cover a broad range of relatively straight forward
designs, and may serve as a template for more complex designs.
}
