% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_and_adjust.R
\name{sample_and_adjust_by_dose}
\alias{sample_and_adjust_by_dose}
\title{Adjust dosing using MIPD on TDMs at specified dose numbers}
\usage{
sample_and_adjust_by_dose(
  adjust_at_dose,
  tdm_times,
  regimen,
  covariates = NULL,
  pars_true_i,
  est_model,
  est_parameters,
  est_omega,
  est_ruv,
  sim_model = NULL,
  sim_ruv = NULL,
  dose_grid = NULL,
  target = list(type = "conc", value = 10),
  target_time = 24,
  ...
)
}
\arguments{
\item{adjust_at_dose}{vector of integers indicating which doses to adjust}

\item{tdm_times}{vector of times to sample at, in hours starting from the
time of first dose.}

\item{regimen}{PKPDsim regimen object, containing initial dosing regimen.}

\item{covariates}{named list of PKPDsim covariates.}

\item{pars_true_i}{PK parameters for the individual. See \code{generate_iiv}.}

\item{est_model}{model to use for estimating patient response.}

\item{est_parameters}{parameters for \code{est_model}.}

\item{est_omega}{omega matrix for \code{est_model}}

\item{est_ruv}{residual variability for \code{est_model}. Named list for
proportional (\code{prop}) and additive (\code{add}) error.}

\item{sim_model}{model to use for simulating "true" patient response. If
NULL, uses \code{est_model}.}

\item{sim_ruv}{residual variability for \code{sim_model}. Named list for
proportional (\code{prop}) and additive (\code{add}) error. If NULL, uses \code{est_ruv}.}

\item{dose_grid}{vector specifying doses to use as test grid for dose
finding, Example: seq(from = 50, to = 500, by = (500 - 50) / 10 ). If
NULL, draws from around the initial dose provided in \code{regimen}.}

\item{target}{target for dose optimization. Named list with \code{type} (e.g.,
"conc", "auc") and \code{value} (e.g., 10).}

\item{target_time}{time (in hours since first dose) at which target should be
estimated for dose finding.}

\item{...}{arguments passed on to \code{simulate_fit} or \code{dose_grid_search}}
}
\value{
a named list containing \code{final_regimen} (all doses after
adjustment), \code{final_estimates} (MAP Bayesian estimation made with all
available levels) and \code{tdms} (all collected levels, both true and measured,
that is, both with and without residual variability).
}
\description{
Doses are iteratively adjusted for the specified dose numbers. Estimation is
performed by MAP Bayesian estimation, and simulation is used to identify the
dose most likely to reach the specified target. See \code{dose_grid_search} for
dose-finding logic. Optionally, disparate models can be used for simulation
and estimation, to allow for intentional model misspecification. Covariates
supplied must support both models. True patient parameters (\code{pars_true_i})
must match the model used for simulation.
}
\details{
Trial design or MIPD protocol can vary considerably in complexity. This
function is intended to cover a broad range of relatively straight forward
designs, and may serve as a template for more complex designs.
}
