% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_subject.R
\name{sim_subject}
\alias{sim_subject}
\title{Core function to simulate a single subject}
\usage{
sim_subject(data, cov_mapping, parameters, design, verbose = FALSE)
}
\arguments{
\item{data}{data for subject}

\item{i}{index number of subject}
}
\description{
Core function to simulate a single subject
}
