% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_subject.R
\name{sim_subject}
\alias{sim_subject}
\title{Core function to simulate a single subject}
\usage{
sim_subject(data, cov_mapping, parameters, design, verbose = FALSE)
}
\arguments{
\item{data}{data for subject}

\item{cov_mapping}{a named vector indicating correspondence between model
covariates and data frame columns. Vector names should reflect all
covariates defined by the model, and vector values should represent the
associated column in the provided data set.}

\item{parameters}{list of model parameters for subject}

\item{design}{trial design, created using \code{create_trial_design()}}

\item{verbose}{verbose output?}
}
\description{
Core function to simulate a single subject
}
