% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosing_nomograms.R
\name{weight_based_starting_dose}
\alias{weight_based_starting_dose}
\title{Weight-based starting dose (e.g., mg/kg)}
\usage{
weight_based_starting_dose(covariates, design, cov_mapping)
}
\description{
This nomogram is a possible \code{method} supplied to
\code{create_initial_regimen_design()}. An argument for \code{mgkg} should also be
supplied to that function call. This function expects that a covariate \code{"WT"}
exists in the processed patient parameters.
}
