% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosing_nomograms.R
\name{weight_based_starting_dose}
\alias{weight_based_starting_dose}
\title{Weight-based starting dose (e.g., mg/kg)}
\usage{
weight_based_starting_dose(covariates, design, cov_mapping)
}
\arguments{
\item{covariates}{named list of PKPDsim covariates. E.g.:
list(WT = PKPDsim::new_covariate(70))}

\item{design}{trial design, created using \code{create_trial_design()}}

\item{cov_mapping}{a named vector indicating correspondence between model
covariates and data frame columns. Vector names should reflect all
covariates defined by the model, and vector values should represent the
associated column in the provided data set.}
}
\description{
This nomogram is a possible \code{method} supplied to
\code{create_initial_regimen_design()}. When this method is used,
\code{create_initial_regimen_design()} must also be supplied with \code{mgkg} and
\code{dosing_weight}, which must match a column name in the supplied data set.
}
