% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosing_nomograms.R
\name{weight_based_starting_dose}
\alias{weight_based_starting_dose}
\title{Weight-based starting dose (e.g., mg/kg)}
\usage{
weight_based_starting_dose(covariates, design, cov_mapping)
}
\description{
This nomogram is a possible \code{method} supplied to
\code{create_initial_regimen_design()}. When this method is used,
\code{create_initial_regimen_design()} must also be supplied with \code{mgkg} and
\code{dosing_weight}, which must match a column name in the supplied data set.
}
